Update thorlabs_MLJ050.py
The call to '._set_home_parameters()' activates the stage to return 'finished home' and 'finished move' messages some time after a call to home/move. The home/move time can vary significantly, so a pause is added (for the estimated time) before requesting the 'finished' message from the serial port.